+++ easycv/models/detection/utils/__init__.py
@@ -2,10 +2,9 @@
 # Copyright (c) 2014-2021 Megvii Inc. All rights reserved.
 from .boxes import (batched_nms, bbox2result, bbox_overlaps, bboxes_iou,
                     box_cxcywh_to_xyxy, box_xyxy_to_cxcywh, distance2bbox,
-                    fp16_clamp, generalized_box_iou, postprocess)
-from .detr_utils import DetrPostProcess
+                    fp16_clamp, generalized_box_iou)
 from .generator import MlvlPointGenerator
 from .misc import (accuracy, filter_scores_and_topk,
                    gen_encoder_output_proposals, gen_sineembed_for_position,
-                   interpolate, inverse_sigmoid, output_postprocess,
-                   select_single_mlvl)
+                   interpolate, inverse_sigmoid, select_single_mlvl)
+from .postprocess import DetrPostProcess, output_postprocess, postprocess
