+++ easycv/models/detection/utils/boxes.py
@@ -36,55 +36,6 @@ def bboxes_iou(bboxes_a, bboxes_b, xyxy=True):
     return area_i / (area_a[:, None] + area_b - area_i)
 
 
-# refer to easycv/models/detection/detectors/yolox/postprocess.py and test.py to rebuild a torch-blade-trtplugin NMS, which is checked by zhoulou in test.py
-# infer docker images is : registry.cn-shanghai.aliyuncs.com/pai-ai-test/eas-service:easycv_blade_181_export
-
-
-def postprocess(prediction, num_classes, conf_thre=0.7, nms_thre=0.45):
-    box_corner = prediction.new(prediction.shape)
-    box_corner[:, :, 0] = prediction[:, :, 0] - prediction[:, :, 2] / 2
-    box_corner[:, :, 1] = prediction[:, :, 1] - prediction[:, :, 3] / 2
-    box_corner[:, :, 2] = prediction[:, :, 0] + prediction[:, :, 2] / 2
-    box_corner[:, :, 3] = prediction[:, :, 1] + prediction[:, :, 3] / 2
-    prediction[:, :, :4] = box_corner[:, :, :4]
-
-    output = [None for _ in range(len(prediction))]
-    for i, image_pred in enumerate(prediction):
-
-        # If none are remaining => process next image
-        if not image_pred.numel():
-            continue
-        # Get score and class with highest confidence
-        class_conf, class_pred = torch.max(
-            image_pred[:, 5:5 + num_classes], 1, keepdim=True)
-
-        conf_mask = (image_pred[:, 4] * class_conf.squeeze() >=
-                     conf_thre).squeeze()
-        # Detections ordered as (x1, y1, x2, y2, obj_conf, class_conf, class_pred)
-        detections = torch.cat(
-            (image_pred[:, :5], class_conf, class_pred.float()), 1)
-        detections = detections[conf_mask]
-        if not detections.numel():
-            continue
-
-        if LooseVersion(torchvision.__version__) >= LooseVersion('0.8.0'):
-            nms_out_index = torchvision.ops.batched_nms(
-                detections[:, :4], detections[:, 4] * detections[:, 5],
-                detections[:, 6], nms_thre)
-        else:
-            nms_out_index = torchvision.ops.nms(
-                detections[:, :4], detections[:, 4] * detections[:, 5],
-                nms_thre)
-
-        detections = detections[nms_out_index]
-        if output[i] is None:
-            output[i] = detections
-        else:
-            output[i] = torch.cat((output[i], detections))
-
-    return output
-
-
 def bbox2result(bboxes, labels, num_classes):
     """Convert detection results to a list of numpy arrays.
     Args:
