+++ easycv/models/detection/utils/misc.py
@@ -137,39 +137,6 @@ def filter_scores_and_topk(scores, score_thr, topk, results=None):
     return scores, labels, keep_idxs, filtered_results
 
 
-def output_postprocess(outputs, img_metas=None):
-    detection_boxes = []
-    detection_scores = []
-    detection_classes = []
-    img_metas_list = []
-
-    for i in range(len(outputs)):
-        if img_metas:
-            img_metas_list.append(img_metas[i])
-        if outputs[i] is not None:
-            bboxes = outputs[i][:, 0:4] if outputs[i] is not None else None
-            if img_metas:
-                bboxes /= img_metas[i]['scale_factor'][0]
-            detection_boxes.append(bboxes.cpu().numpy())
-            detection_scores.append(
-                (outputs[i][:, 4] * outputs[i][:, 5]).cpu().numpy())
-            detection_classes.append(outputs[i][:, 6].cpu().numpy().astype(
-                np.int32))
-        else:
-            detection_boxes.append(None)
-            detection_scores.append(None)
-            detection_classes.append(None)
-
-    test_outputs = {
-        'detection_boxes': detection_boxes,
-        'detection_scores': detection_scores,
-        'detection_classes': detection_classes,
-        'img_metas': img_metas_list
-    }
-
-    return test_outputs
-
-
 def inverse_sigmoid(x, eps=1e-3):
     x = x.clamp(min=0, max=1)
     x1 = x.clamp(min=eps)
